backend/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==4.2.6
django-cache-url==3.4.4
django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
environs==9.5.0
gunicorn==21.2.0
marshmallow==3.20.1
packaging==23.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
